+++ pkg/local_object_storage/shard/delete.go
@@ -82,6 +82,10 @@ func (s *Shard) delete(prm DeletePrm) (DeleteRes, error) {
 	s.decObjectCounterBy(logical, res.AvailableObjectsRemoved())
 
 	for i := range prm.addr {
+		s.addToContainerSize(prm.addr[i].Container().EncodeToString(), -int64(res.RemovedObjectSizes()[i]))
+	}
+
+	for i := range prm.addr {
 		var delPrm common.DeletePrm
 		delPrm.Address = prm.addr[i]
 		id := smalls[prm.addr[i]]